_execute() requires the declaration of $params
As can be seen in line 281, the execute() method is calling _execute() with the validated $options obtained from CLI.

We must declare this argument in the abstract function so that overriding Task classes are able to receive this array.

```
// Finally, run the task
$this->{$this->_method}($options);
```

The implementation in Minion has always been that "$this->_options" represents the list of available parameters with their default values. In runtime, "$params" represents the validated input from the user, or simply the default values in lieu of any input.

I don't see a reason to introduce an API breaking change as part of this refactor.